Add Explicit Package Names to its Functions

Add Explicit Package Names to its Functions

originize_text(
  text,
  pkgs = getOption("origin.pkgs", .packages()),
  overwrite = getOption("origin.overwrite", TRUE),
  ask_before_applying_changes = getOption("origin.ask_before_applying_changes", TRUE),
  ignore_comments = getOption("origin.ignore_comments", TRUE),
  check_conflicts = getOption("origin.check_conflicts", TRUE),
  check_base_conflicts = getOption("origin.check_base_conflicts", TRUE),
  add_base_packages = getOption("origin.add_base_packages", FALSE),
  excluded_functions = getOption("origin.excluded_functions", list()),
  verbose = getOption("origin.verbose", FALSE),
  use_markers = getOption("origin.use_markers_for_logging", TRUE)
)

Arguments

text
string to originize

pkgs
a character vector with package names

overwrite
a boolean, if TRUE the file will be saved and overwritten. If FALSE the file is returned.

ask_before_applying_changes
a boolean. If TRUE, the user has to approve changes made by origin prior to applying them

ignore_comments
a boolean, if TRUE lines starting with # are ignored

check_conflicts
a boolean. If TRUE, possible namespace conflicts are checked and, if there are any, user input is required. See also

check_base_conflicts
a boolean. If TRUE; base R functions are also included in checking for conflicts

add_base_packages
a boolean. If TRUE, base R functions are handled like all other packages and added via `::`

excluded_functions
a list. Either an unnamed list of function names as strings. Then these functions are excluded from all packages and never considered in origin. Or a named list with character vectors, Then the name of the list element refers to a package and the given functions are only excluded from this package. A very explicit way to handle namespace conflicts

verbose
a boolean. If TRUE, origin provides feedback about its steps.

use_markers
a boolean. If TRUE, the markers tab inn RStudio is used to track changes and show issues. FALSE prints the same information in the console.